Another couple of Windows fixes.

+++ fmpz_lll/is_reduced_mpfr.c
@@ -123,7 +123,7 @@ fmpz_lll_is_reduced_mpfr(const fmpz_mat_t B, const fmpz_lll_t fl,
                              norm, MPFR_RNDN);
                 }
                 mpfr_mul_si(mpfr_mat_entry(V, j, i), mpfr_mat_entry(V, j, i),
-                            -1, MPFR_RNDN);
+                            -WORD(1), MPFR_RNDN);
                 mpfr_mul(mpfr_mat_entry(V, j, i), mpfr_mat_entry(V, j, i),
                          mpfr_mat_entry(V, j, j), MPFR_RNDN);
             }
@@ -591,7 +591,7 @@ fmpz_lll_is_reduced_mpfr(const fmpz_mat_t B, const fmpz_lll_t fl,
                              norm, MPFR_RNDN);
                 }
                 mpfr_mul_si(mpfr_mat_entry(V, j, i), mpfr_mat_entry(V, j, i),
-                            -1, MPFR_RNDN);
+                            -WORD(1), MPFR_RNDN);
                 mpfr_mul(mpfr_mat_entry(V, j, i), mpfr_mat_entry(V, j, i),
                          mpfr_mat_entry(V, j, j), MPFR_RNDN);
             }
