+++ fmpz_lll/is_reduced_mpfr_with_removal.c
@@ -124,7 +124,7 @@ fmpz_lll_is_reduced_mpfr_with_removal(const fmpz_mat_t B, const fmpz_lll_t fl,
                              norm, MPFR_RNDN);
                 }
                 mpfr_mul_si(mpfr_mat_entry(V, j, i), mpfr_mat_entry(V, j, i),
-                            -1, MPFR_RNDN);
+                            -WORD(1), MPFR_RNDN);
                 mpfr_mul(mpfr_mat_entry(V, j, i), mpfr_mat_entry(V, j, i),
                          mpfr_mat_entry(V, j, j), MPFR_RNDN);
             }
@@ -612,7 +612,7 @@ fmpz_lll_is_reduced_mpfr_with_removal(const fmpz_mat_t B, const fmpz_lll_t fl,
                              norm, MPFR_RNDN);
                 }
                 mpfr_mul_si(mpfr_mat_entry(V, j, i), mpfr_mat_entry(V, j, i),
-                            -1, MPFR_RNDN);
+                            -WORD(1), MPFR_RNDN);
                 mpfr_mul(mpfr_mat_entry(V, j, i), mpfr_mat_entry(V, j, i),
                          mpfr_mat_entry(V, j, j), MPFR_RNDN);
             }
